PM-11240: Moving models out of entities namespace
