+++ src/Core/ActionableInsights/Models/ApplicationReportParameters.cs
@@ -1,0 +1,9 @@
+﻿#nullable enable
+namespace Bit.Core.ActionableInsights.Models;
+
+public class ApplicationReportParameters
+{
+    public string Name { get; set; } = null!;
+
+    public required IEnumerable<UrisReportParameters> Uris { get; set; }
+}
