+++ src/Core/ActionableInsights/Models/ReportParameters.cs
@@ -1,0 +1,7 @@
+﻿#nullable enable
+namespace Bit.Core.ActionableInsights.Models;
+
+public class ReportParameters
+{
+    public IEnumerable<ApplicationReportParameters>? Applications { get; set; }
+}
